Moved all editing/ownership functionality for chatroom into Chat element
